Update README.md
Geometry - area of a polygon

Part 1: Area of a hexagon

The area of a hexagon can be computed using the following formula (s is the length of a side):



Write a program that prompts the user to enter the side of a hexagon and displays its area. Here is a sample run:

Enter the side: 5.5
The area of the hexagon is 78.59

Part 2: Area of a regular polygon

A regular polygon is an n-sided polygon in which all sides are of the same length and all angles have the same degree (i.e., the polygon is both equilateral and equiangular). The formula for computing the area of a regular polygon is



Here, s is the length of a side.

Modify your program from Part 1 so that it will now find the area of a regular polygon. The program should prompt the user to enter the number of sides and their length for a regular polygon and displays its area. Here is a sample run:

Enter the number of sides: 5
Enter the side: 6.5
The area of the polygon is 72.69